+++ examples/basic/nasti.config.ts
@@ -1,16 +1,16 @@
 import { defineConfig } from '@nasti-toolchain/nasti'
 import { pluginVueLynx } from '@nasti-toolchain/plugin-vue-lynx'
 
-const useNativeBackend = process.env['VUE_LYNX_NATIVE_BACKEND'] === '1'
+const useRspeedyBackend = process.env['VUE_LYNX_RSPEEDY_BACKEND'] === '1'
 
 export default defineConfig({
   plugins: [
     pluginVueLynx({
-      backend: useNativeBackend ? 'nasti' : 'rspeedy',
+      backend: useRspeedyBackend ? 'rspeedy' : 'nasti',
       entry: {
         main: './src/index.ts',
       },
-      web: useNativeBackend ? false : true,
+      web: useRspeedyBackend ? true : false,
       vue: {
         optionsApi: false,
       },